Move AppFooter component to matters dir
