+++ layouts/default.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import AppFooter from '~/components/AppFooter'
+import AppFooter from '~/components/matters/AppFooter'
 
 export default {
   components: {
